adding mp4 movie creation

+++ openfoamDNS_to_pseudoPIV_all.csh
@@ -974,6 +974,13 @@ foreach CASE ( ${All_CASE} )
         
         ### creating movies from PNG files (if more than [min_Nb_t] time directories)
         
+        # flv : bad quality
+        # avi : good quality but bigger than MP4 file
+        # mp4 : good quality
+        set movie_type = flv 
+        set movie_type = avi 
+        set movie_type = mp4 
+        
         ## temporary movie folder : PNG files from D folder are renamed and copied in temporary movie folder,
         #                                         if there is more than min_Nb_t (=5) time directories (i.e. when D=residualSpeed_*)
         #                                         (for example [residualSpeed_2/750.25/PIV_new_Ux.png] is copied in [tmp_movie/75025.png])
@@ -993,30 +1000,54 @@ foreach CASE ( ${All_CASE} )
             PIV_new_Uy.png \
           )
           foreach png_file ( ${All_png} )
-            
-            set flv_file = ` echo ${png_file} | sed s/"png"/"flv"/g `
+                        
+            cd ${dir_work_up}
 
+            set flv_file = ` echo ${png_file} | sed s/"png"/"flv"/g `
+            set avi_file = ` echo ${png_file} | sed s/"png"/"avi"/g `
+            set mp4_file = ` echo ${png_file} | sed s/"png"/"mp4"/g `
+
+            set i = 1
             foreach t ( ${All_t} )
 
-              set nom = `echo ${t} | awk '{ printf("%06.0f",$1*100.) }' `
+              set nom = `echo ${i} | awk '{ printf("%04d",$1) }' `
               if -e ${dir_work}/${t}/${png_file} \cp ${dir_work}/${t}/${png_file} tmp_movie/${nom}.png
+              
+              set i = ` echo ${i} | awk '{ print $1+1}' `
               
             end
             
-            # creating AVI movie
+            if ( ${movie_type} == "mp4" ) then
             
-            mencoder "mf://tmp_movie/??????.png" -o frame.avi -ovc lavc -lavcopts vcodec=mjpeg
-            #mplayer -speed 0.1 frame.avi
+              ffmpeg -r 10 -i tmp_movie/%04d.png ${dir_work}/${mp4_file}
+              set info = "Cf. file ${dir_work}/${mp4_file}\n (mplayer -speed 0.5 ${dir_work}/${mp4_file}) "
+           
+            else
+           
+              # creating AVI movie
+              
+              mencoder "mf://tmp_movie/????.png" -o frame.avi -ovc lavc -lavcopts vcodec=mjpeg
+              #mplayer -speed 0.1 frame.avi
 
-            # converting AVI to FLV
-            
-            ffmpeg -i frame.avi -f flv -y frame.flv
-            if -e frame.avi \rm frame.avi
-            #vlc --rate 0.5 frame.flv
-            # mplayer -speed 0.5 frame.flv
-            if -e frame.flv \mv frame.flv ${dir_work}/${flv_file}
-            
-            set info = "Cf. file ${dir_work}/${flv_file}\n (mplayer -speed 0.5 ${dir_work}/${flv_file}) "
+              if  ( ${movie_type} == "avi" ) then
+              
+                mv frame.avi ${dir_work}/${avi_file}
+                set info = "Cf. file ${dir_work}/${avi_file}\n (mplayer -speed 0.5 ${dir_work}/${avi_file}) "
+              
+              else
+              
+                # converting AVI to FLV
+                
+                ffmpeg -i frame.avi -f flv -y frame.flv
+                if -e frame.avi \rm frame.avi
+                #vlc --rate 0.5 frame.flv
+                # mplayer -speed 0.5 frame.flv
+                if -e frame.flv \mv frame.flv ${dir_work}/${flv_file}
+                
+                set info = "Cf. file ${dir_work}/${flv_file}\n (mplayer -speed 0.5 ${dir_work}/${flv_file}) "
+          
+            endif
+
             WRITE_LINE; WRITE_INFO "${info}";  WRITE_LINE
           
           end
